label: store 'Rect_t' boundaries as doubles
The boundaries remain round numbers, but this change addresses a problem wherein
the range of expressible boundaries was limited to [`INT_MIN`, `INT_MAX`].
Boundaries can now span the full range of `double`, [-2⁵³, 2⁵³].

This squashes 8 -Wfloat-conversion warnings, 4 -Wbad-function-cast warnings, and
1 -Wsign-conversion warning.

+++ lib/label/rectangle.h
@@ -18,7 +18,7 @@ extern "C" {
 #endif
 
 typedef struct Rect {
-    int boundary[NUMSIDES];
+  double boundary[NUMSIDES];
 } Rect_t;
 
 void InitRect(Rect_t * r);
